Front-End Improvements.
Front-End no longer uses shadow-roots, improving cross browser compatibility. The Buttons have also been removed and been replaced with onclick listeners on the links that have been added to the individual files to improve accuracy (not having to go down the long row).

Large files (like SWFs) are no longer attached to the emails, instead a direct link is provided for the requestor to directly download the file.

Added a FavIcon as well as various other CSS improvements.